Initial commit of the PCS storage layer.
This layer contains an interface for PCS storage operations, storage
being an external data store that is independent of all PCS instances.
This layer currently includes an ETCD and an in-memory implementation,
both based on hms-hmetcd.

In the future we may also add support for memcached.

This was tested locally, via new unit tests, and via the PCS integration
tests.   Note that hms-hmetcd has been tested to death, so the only
real testing was to insure our calls to hms-hmetcd work correctly.