Fixed return value of interface method, fixed RTS issues.
Interface method return value was incorrectly specified in the
interface definition.   This was fixed.

In the integrated unit test, RTS was not using an HTTPS transport.
This causes issues since RTS hard codes the FQDN of any simulated
BMC as xname:8083, which must use the HTTPS transport.  If HTTP
is used, it will use port 8082 and nothing will work.  This may
need to be addressed in RTS, but for now, .crt and .key files
were re-introduced to make RTS use HTTPS and work correctly.